pyproject.toml
[tool.poetry]
name = "StockWatch"
version = "0.1.0"
description = "A FastAPI-powered stock market API that fetches real-time stock prices"
authors = ["Apeksha Gaonkar <gaonkarapeksha1234@gmail.com>"]
readme = "README.md"
# TODO: Rename the package/directory.
packages = [{include = "StockWatch", from = "src"}]

[tool.poetry.dependencies]
python = "~3.12.1"
fastapi = "^0.115.11"
uvicorn = "^0.30.3"
requests = "^2.32.3"
starlette = "^0.46.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
python-dotenv = "^1.0.1"
pytest-cov = "^6.0.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.1.2"
bandit = "^1.8.3"
safety = "^3.3.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
